requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
openai-whisper==20231117
deep-translator==1.11.4
langdetect==1.0.9